+++ test/support/domain.ex
@@ -244,6 +244,7 @@ defmodule AshHateoas.Test.Domain do
     resource(AshHateoas.Test.NoBase)
     resource(AshHateoas.Test.Compensating)
     resource(AshHateoas.Test.MultiRead)
+    resource(AshHateoas.Test.EagerPrepare)
     resource(AshHateoas.Test.AuthUser)
     resource(AshHateoas.Test.AuthToken)
   end